+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ x.x.x (x-x-x)
 * Improve exception handling.
 * Removed support for Python 3.6 (end of life since December 2021).
 * Added support for Python 3.10 and Python 3.11.
+* Added logout command.
 
 3.17.2 (2023-03-16)
 -------------------